[SPARK-30010] Remove deprecated setAll(Traversable) overload
This overload became ambiguous in Scala 2.13.

This was previously proposed as
https://github.com/apache/spark/pull/27308 but rejected because the
overload had not been deprecated in a released version of Spark yet. One
way to get Spark 3.0 on Scala 2.13 anyway would be to have separate 2.12
and 2.13 versions of the SparkConf.scala source file and only remove the
overload in the 2.13 version.